Finalização

+++ gamagit/src/pages/Repositories/index.js
@@ -1,13 +1,21 @@
 import React, { useEffect, useState } from 'react';
-import * as S from './styled'
+import * as S from './styled';
+import { useHistory } from 'react-router-dom';
 
 export default function Repositories() {
-    const [ repositories, setRepositories ] = useState([])
+    const history = useHistory();
+    const [ repositories, setRepositories ] = useState([]);
+
     useEffect(() => {
         let repositoriesName = localStorage.getItem('repositoriesName');
-        repositoriesName = JSON.parse(repositoriesName);
-        setRepositories(repositoriesName);
-        localStorage.clear();
+        if(repositoriesName != null){
+            repositoriesName = JSON.parse(repositoriesName);
+            setRepositories(repositoriesName);
+            localStorage.clear();
+        } else {
+            history.push('/');
+        }
+        
     }, []);
 
     return (
@@ -20,6 +28,7 @@ export default function Repositories() {
                     )
                 }) }
             </S.List>
+            <S.LinkHome to="/">Voltar</S.LinkHome>
         </S.Container>
     )
 }
